Rollback Kanon class  to __init__ module
